refactored code

+++ src/pathtrace.h
@@ -10,4 +10,4 @@ void showGBuffer(uchar4 *pbo);
 void showImage(uchar4 *pbo, int iter);
 void showDenoise(uchar4 *pbo, int iter);
 
-bool DenoiseImage(int resolutionX, int resolutionY, int iteration, int filterSize, int colWeight, int norWeight, int posWeight);
+bool DenoiseImage(int resolutionX, int resolutionY, int iteration, int filterSize, float colWeight, float norWeight, float posWeight);
